aula10

+++ aula10 - integrando com o banco de dados/controller/controller_aluno.js
@@ -21,59 +21,71 @@ const getAlunos = async () => {
 
     if (dadosAlunos) {
         //Criando um JSON com o atributo alunos, para encaminhar um ARRAY de alunos
+        dadosAlunosJSON.status = message.SUCCESS_REQUEST.status
+        dadosAlunosJSON.message = message.SUCCESS_REQUEST.message
         dadosAlunosJSON.quantidade = dadosAlunos.length
         dadosAlunosJSON.alunos = dadosAlunos
         return dadosAlunosJSON
     } else {
-        return false
+        return message.ERROR_NOT_FOUND
     }
-
 }
 
 //retorna o aluno filtrando pelo ID
-const getBuscarAlunoId = async (id) => {
-    let idAluno = id
+const getBuscarAlunoId = async (idAluno) => {
+
     let dadosAlunosJSON = {}
 
+    //validação do idAluno
     if (idAluno == "" || idAluno == undefined || isNaN(idAluno)) {
         return message.ERROR_INVALID_ID
     } else {
+
         //Chama a função do arquivo DAO que irá retorna os alunos pelo id
-        let dadosAlunos = await alunoDAO.selectByIdAluno(idAluno)
+        let dadosAlunos = await alunoDAO.selectAlunoById(idAluno)
 
         if (dadosAlunos) {
-            dadosAlunosJSON.quantidade = dadosAlunos.length
+            dadosAlunosJSON.status = message.SUCCESS_REQUEST.status
+            dadosAlunosJSON.message = message.SUCCESS_REQUEST.message
             dadosAlunosJSON.aluno = dadosAlunos
             return dadosAlunosJSON
         } else {
-            return false
+            return message.ERROR_NOT_FOUND
         }
     }
-}
 
-const getBuscarAlunoNome = async (nome) => {
-    let nomeAluno = nome
+}
+//retorna o aluno filtrando pelo ID
+
+//retorna o aluno filtrando pelo nome
+const getBuscarAlunoNome = async (nomeAluno) => {
+
     let dadosAlunosJSON = {}
 
-    if (nomeAluno == "" || nomeAluno == undefined || isNaN(nomeAluno)) {
-        return message.ERROR_INVALID_ID
+    if (nomeAluno == "" || nomeAluno == undefined || !isNaN(nomeAluno)) {
+        return message.ERROR_INVALID_NAME
     } else {
-        //Chama a função do aqruivo DAO que irá retorna todos os alunos pelo nome
-        let dadosAlunos = await alunoDAO.selectByNameAluno(nomeAluno)
+
+        //chama a função do arquivo DAO que irá retorna um aluno pelo nome
+        let dadosAlunos = await alunoDAO.selectAlunoByName(nomeAluno)
 
         if (dadosAlunos) {
-            dadosAlunosJSON.quantidade = dadosAlunos.length
+
+            dadosAlunosJSON.status = message.SUCCESS_REQUEST.status
+            dadosAlunosJSON.message = message.SUCCESS_REQUEST.message
             dadosAlunosJSON.aluno = dadosAlunos
+
             return dadosAlunosJSON
+
         } else {
-            return false
+            return message.ERROR_NOT_FOUND
         }
     }
+
 }
 
 //inserir um novo aluno
 const novoAluno = async (dadosAluno) => {
-    let resultDadosAluno
 
     //Validação para tratar campos obrigatórios e quantidade de caracteres
     if (dadosAluno.nome == '' || dadosAluno.nome == undefined || dadosAluno.nome.length > 100 ||
@@ -84,21 +96,33 @@ const novoAluno = async (dadosAluno) => {
     ) {
         return message.ERROR_REQUIRED_FIELDS
     } else {
+
         //Envia os dados para a model inserir no banco de dados
-        resultDadosAluno = await alunoDAO.insertAluno(dadosAluno)
+        let resultDadosAluno = await alunoDAO.insertAluno(dadosAluno)
 
         //valida se o banco de dados inseriu corretamente os dados
         if (resultDadosAluno) {
-            return message.SUCCESS_CREATE_ITEM
+
+            //chama a função que vai encontar o ID gerado após o insert
+            let novoAluno = await alunoDAO.selectLastId()
+
+            let dadosAlunosJSON = {}
+
+            dadosAlunosJSON.status = message.SUCCESS_CREATE_ITEM.status
+            dadosAlunosJSON.message = message.SUCCESS_CREATE_ITEM.message
+            dadosAlunosJSON.aluno = novoAluno
+
+            return dadosAlunosJSON
+
         } else {
             return message.ERROR_INTERNAL_SERVER
         }
     }
+    
 }
 
 //atualiza um aluno existente
 const atualizarAluno = async (dadosAluno, idAluno) => {
-    let resultDadosAluno
 
     //Validação para tratar campos obrigatórios e quantidade de caracteres
     if (dadosAluno.nome == '' || dadosAluno.nome == undefined || dadosAluno.nome.length > 100 ||
@@ -107,40 +131,75 @@ const atualizarAluno = async (dadosAluno, idAluno) => {
         dadosAluno.data_nascimento == '' || dadosAluno.data_nascimento == undefined || dadosAluno.data_nascimento.length > 10 ||
         dadosAluno.email == '' || dadosAluno.email == undefined || dadosAluno.email.length > 255
     ) {
+
         return message.ERROR_REQUIRED_FIELDS //status code 400
-        //Validação de ID incorreto ou não informado
+
+        //validação de ID incorreto ou não informado
     } else if (idAluno == '' || idAluno == undefined || isNaN(idAluno)) {
+
         return message.ERROR_INVALID_ID //status code 400
+
     } else {
+
+        //adiciona o ID  do aluno no json dos dadosAluno
         dadosAluno.id = idAluno
 
-        //encaminha os dados para a model do aluno
-        resultDadosAluno = await alunoDAO.updateAluno(dadosAluno)
+        let statusId = await alunoDAO.selectAlunoById(idAluno)
 
-        if (resultDadosAluno) {
-            return message.SUCCESS_UPDATE_ITEM //200
+        //validação para ver se o ID realmente existe
+        if (statusId) {
+
+            //chama a função que vai atualizar o aluno
+            let resultDadosAluno = await alunoDAO.updateAluno(dadosAluno)
+
+            if (resultDadosAluno) {
+
+                let dadosAlunosJSON = {}
+
+                dadosAlunosJSON.status = message.SUCCESS_UPDATE_ITEM.status //200
+                dadosAlunosJSON.message = message.SUCCESS_UPDATE_ITEM.message
+                dadosAlunosJSON.aluno = dadosAluno
+
+                return dadosAlunosJSON
+
+            } else {
+                return message.ERROR_INTERNAL_SERVER //500
+            }
+
         } else {
-            return message.ERROR_INTERNAL_SERVER
+            return message.ERROR_ID_NOT_FOUND //404
         }
     }
+
 }
 
 //deleta um novo aluno existente
-const deletarAluno = async (id) => {
-    let idAluno = id
-    let resultDadosAluno
+const deletarAluno = async (idAluno) => {
 
     if (idAluno == "" || idAluno == undefined || isNaN(idAluno)) {
         return message.ERROR_INVALID_ID
     } else {
-        resultDadosAluno = await alunoDAO.deleteAluno(idAluno)
 
-        if (resultDadosAluno) {
-            return message.SUCCESS_DELETE_ITEM
+        //chama a função para ver o ID do aluno
+        let statusId = await alunoDAO.selectAlunoById(idAluno)
+
+        //validação para ver se o ID realmente existe
+        if (statusId) {
+
+            //chama a função que vai apagar o aluno
+            let resultDadosAluno = await alunoDAO.deleteAluno(idAluno)
+
+            if (resultDadosAluno) {
+                return message.SUCCESS_DELETE_ITEM
+            } else {
+                return message.ERROR_INTERNAL_SERVER
+            }
+
         } else {
-            return message.ERROR_INTERNAL_SERVER
+            return message.ERROR_ID_NOT_FOUND
         }
     }
+
 }
 
 module.exports = {
